@@ -85,7 +85,6 @@
     "eslint-plugin-prettier": "^4.2.1",
     "fixturify": "^2.1.1",
     "fs-extra": "^10.1.0",
-    "get-package-type": "^0.1.0",
     "github-api": "^3.4.0",
     "hash.js": "^1.1.7",
     "husky": "^8.0.1",
@@ -125,7 +124,7 @@
     "dist/es/package.json"
   ],
   "engines": {
-    "node": ">=14.13.1",
+    "node": ">=14.18.0",
     "npm": ">=8.0.0"
   },
   "exports": {
