@@ -1,10 +1,17 @@
-import { extname, isAbsolute } from 'node:path';
+import { promises as fs } from 'node:fs';
+import { dirname, isAbsolute, join } from 'node:path';
+import process from 'node:process';
 import { pathToFileURL } from 'node:url';
-import getPackageType from 'get-package-type';
 import * as rollup from '../../src/node-entry';
 import type { MergedRollupOptions } from '../../src/rollup/types';
 import { bold } from '../../src/utils/colors';
-import { errMissingConfig, error, errTranspiledEsmConfig } from '../../src/utils/error';
+import {
+	errCannotBundleConfigAsEsm,
+	errCannotLoadConfigAsCjs,
+	errCannotLoadConfigAsEsm,
+	errMissingConfig,
+	error
+} from '../../src/utils/error';
 import { mergeOptions } from '../../src/utils/options/mergeOptions';
 import type { GenericConfigObject } from '../../src/utils/options/options';
 import relativeId from '../../src/utils/relativeId';
@@ -12,15 +19,14 @@ import { stderr } from '../logging';
 import batchWarnings, { type BatchWarnings } from './batchWarnings';
 import { addCommandPluginsToInputOptions, addPluginsFromCommandOption } from './commandPlugins';
 
-interface NodeModuleWithCompile extends NodeModule {
-	_compile(code: string, filename: string): any;
-}
-
 export async function loadConfigFile(
 	fileName: string,
 	commandOptions: any = {}
 ): Promise<{ options: MergedRollupOptions[]; warnings: BatchWarnings }> {
-	const configs = await loadConfigsFromFile(fileName, commandOptions);
+	const configs = await getConfigList(
+		getDefaultFromCjs(await getConfigFileExport(fileName, commandOptions)),
+		commandOptions
+	);
 	const warnings = batchWarnings();
 	try {
 		const normalizedConfigs: MergedRollupOptions[] = [];
@@ -36,29 +42,51 @@ export async function loadConfigFile(
 	}
 }
 
-async function loadConfigsFromFile(
-	fileName: string,
-	commandOptions: Record<string, unknown>
-): Promise<GenericConfigObject[]> {
-	const extension = extname(fileName);
-
-	const configFileExport =
-		commandOptions.configPlugin ||
-		// We always transpile the .js non-module case because many legacy code bases rely on this
-		(extension === '.js' && getPackageType.sync(fileName) !== 'module')
-			? await getDefaultFromTranspiledConfigFile(fileName, commandOptions)
-			: getDefaultFromCjs((await import(pathToFileURL(fileName).href)).default);
-
-	return getConfigList(configFileExport, commandOptions);
+async function getConfigFileExport(fileName: string, commandOptions: Record<string, unknown>) {
+	if (commandOptions.configPlugin || commandOptions.bundleConfigAsCjs) {
+		try {
+			return await loadTranspiledConfigFile(fileName, commandOptions);
+		} catch (err: any) {
+			if (err.message.includes('not defined in ES module scope')) {
+				return error(errCannotBundleConfigAsEsm(err));
+			}
+			throw err;
+		}
+	}
+	let cannotLoadEsm = false;
+	const handleWarning = (warning: Error): void => {
+		if (warning.message.includes('To load an ES module')) {
+			cannotLoadEsm = true;
+		}
+	};
+	process.on('warning', handleWarning);
+	try {
+		const fileUrl = pathToFileURL(fileName);
+		if (process.env.ROLLUP_WATCH) {
+			// We are adding the current date to allow reloads in watch mode
+			fileUrl.search = `?${Date.now()}`;
+		}
+		return (await import(fileUrl.href)).default;
+	} catch (err: any) {
+		if (cannotLoadEsm) {
+			return error(errCannotLoadConfigAsCjs(err));
+		}
+		if (err.message.includes('not defined in ES module scope')) {
+			return error(errCannotLoadConfigAsEsm(err));
+		}
+		throw err;
+	} finally {
+		process.off('warning', handleWarning);
+	}
 }
 
 function getDefaultFromCjs(namespace: GenericConfigObject): unknown {
-	return namespace.__esModule ? namespace.default : namespace;
+	return namespace.default || namespace;
 }
 
-async function getDefaultFromTranspiledConfigFile(
+async function loadTranspiledConfigFile(
 	fileName: string,
-	commandOptions: Record<string, unknown>
+	{ bundleConfigAsCjs, configPlugin, silent }: Record<string, unknown>
 ): Promise<unknown> {
 	const warnings = batchWarnings();
 	const inputOptions = {
@@ -69,9 +97,9 @@ async function getDefaultFromTranspiledConfigFile(
 		plugins: [],
 		treeshake: false
 	};
-	await addPluginsFromCommandOption(commandOptions.configPlugin, inputOptions);
+	await addPluginsFromCommandOption(configPlugin, inputOptions);
 	const bundle = await rollup.rollup(inputOptions);
-	if (!commandOptions.silent && warnings.count > 0) {
+	if (!silent && warnings.count > 0) {
 		stderr(bold(`loaded ${relativeId(fileName)} with warnings`));
 		warnings.flush();
 	}
@@ -79,7 +107,7 @@ async function getDefaultFromTranspiledConfigFile(
 		output: [{ code }]
 	} = await bundle.generate({
 		exports: 'named',
-		format: 'cjs',
+		format: bundleConfigAsCjs ? 'cjs' : 'es',
 		plugins: [
 			{
 				name: 'transpile-import-meta',
@@ -94,32 +122,22 @@ async function getDefaultFromTranspiledConfigFile(
 			}
 		]
 	});
-	return loadConfigFromBundledFile(fileName, code);
+	return loadConfigFromWrittenFile(
+		join(dirname(fileName), `rollup.config-${Date.now()}.${bundleConfigAsCjs ? 'cjs' : 'mjs'}`),
+		code
+	);
 }
 
-function loadConfigFromBundledFile(fileName: string, bundledCode: string): unknown {
-	const resolvedFileName = require.resolve(fileName);
-	const extension = extname(resolvedFileName);
-	const defaultLoader = require.extensions[extension];
-	require.extensions[extension] = (module: NodeModule, requiredFileName: string) => {
-		if (requiredFileName === resolvedFileName) {
-			(module as NodeModuleWithCompile)._compile(bundledCode, requiredFileName);
-		} else {
-			if (defaultLoader) {
-				defaultLoader(module, requiredFileName);
-			}
-		}
-	};
-	delete require.cache[resolvedFileName];
+async function loadConfigFromWrittenFile(
+	bundledFileName: string,
+	bundledCode: string
+): Promise<unknown> {
+	await fs.writeFile(bundledFileName, bundledCode);
 	try {
-		const config = getDefaultFromCjs(require(fileName));
-		require.extensions[extension] = defaultLoader;
-		return config;
-	} catch (err: any) {
-		if (err.code === 'ERR_REQUIRE_ESM') {
-			return error(errTranspiledEsmConfig(fileName));
-		}
-		throw err;
+		return (await import(pathToFileURL(bundledFileName).href)).default;
+	} finally {
+		// Not awaiting here saves some ms while potentially hiding a non-critical error
+		fs.unlink(bundledFileName);
 	}
 }
 
