@@ -71,6 +71,7 @@ const ADDON_ERROR = 'ADDON_ERROR',
 	ILLEGAL_REASSIGNMENT = 'ILLEGAL_REASSIGNMENT',
 	INPUT_HOOK_IN_OUTPUT_PLUGIN = 'INPUT_HOOK_IN_OUTPUT_PLUGIN',
 	INVALID_CHUNK = 'INVALID_CHUNK',
+	INVALID_CONFIG_MODULE_FORMAT = 'INVALID_CONFIG_MODULE_FORMAT',
 	INVALID_EXPORT_OPTION = 'INVALID_EXPORT_OPTION',
 	INVALID_EXTERNAL_ID = 'INVALID_EXTERNAL_ID',
 	INVALID_OPTION = 'INVALID_OPTION',
@@ -100,7 +101,6 @@ const ADDON_ERROR = 'ADDON_ERROR',
 	SOURCEMAP_ERROR = 'SOURCEMAP_ERROR',
 	SYNTHETIC_NAMED_EXPORTS_NEED_NAMESPACE_EXPORT = 'SYNTHETIC_NAMED_EXPORTS_NEED_NAMESPACE_EXPORT',
 	THIS_IS_UNDEFINED = 'THIS_IS_UNDEFINED',
-	TRANSPILED_ESM_CONFIG = 'TRANSPILED_ESM_CONFIG',
 	UNEXPECTED_NAMED_IMPORT = 'UNEXPECTED_NAMED_IMPORT',
 	UNKNOWN_OPTION = 'UNKNOWN_OPTION',
 	UNRESOLVED_ENTRY = 'UNRESOLVED_ENTRY',
@@ -362,6 +362,36 @@ export function errCannotAssignModuleToChunk(
 		message: `Cannot assign "${relativeId(
 			moduleId
 		)}" to the "${assignToAlias}" chunk as it is already in the "${currentAlias}" chunk.`
+	};
+}
+
+export function errCannotBundleConfigAsEsm(originalError: Error): RollupLog {
+	return {
+		cause: originalError,
+		code: INVALID_CONFIG_MODULE_FORMAT,
+		message: `Rollup transpiled your configuration to an  ES module even though it appears to contain CommonJS elements. To resolve this, you can pass the "--bundleConfigAsCjs" flag to Rollup or change your configuration to only contain valid ESM code.\n\nOriginal error: ${originalError.message}`,
+		stack: originalError.stack,
+		url: 'https://rollupjs.org/guide/en/#--bundleconfigascjs'
+	};
+}
+
+export function errCannotLoadConfigAsCjs(originalError: Error): RollupLog {
+	return {
+		cause: originalError,
+		code: INVALID_CONFIG_MODULE_FORMAT,
+		message: `Node tried to load your configuration file as CommonJS even though it is likely an ES module. To resolve this, change the extension of your configuration to ".mjs", set "type": "module" in your package.json file or pass the "--bundleConfigAsCjs" flag.\n\nOriginal error: ${originalError.message}`,
+		stack: originalError.stack,
+		url: 'https://rollupjs.org/guide/en/#--bundleconfigascjs'
+	};
+}
+
+export function errCannotLoadConfigAsEsm(originalError: Error): RollupLog {
+	return {
+		cause: originalError,
+		code: INVALID_CONFIG_MODULE_FORMAT,
+		message: `Node tried to load your configuration as an ES module even though it is likely CommonJS. To resolve this, change the extension of your configuration to ".cjs" or pass the "--bundleConfigAsCjs" flag.\n\nOriginal error: ${originalError.message}`,
+		stack: originalError.stack,
+		url: 'https://rollupjs.org/guide/en/#--bundleconfigascjs'
 	};
 }
 
@@ -753,15 +783,6 @@ export function errThisIsUndefined(): RollupLog {
 		code: THIS_IS_UNDEFINED,
 		message: `The 'this' keyword is equivalent to 'undefined' at the top level of an ES module, and has been rewritten`,
 		url: `https://rollupjs.org/guide/en/#error-this-is-undefined`
-	};
-}
-
-export function errTranspiledEsmConfig(fileName: string): RollupLog {
-	return {
-		code: TRANSPILED_ESM_CONFIG,
-		message: `While loading the Rollup configuration from "${relativeId(
-			fileName
-		)}", Node tried to require an ES module from a CommonJS file, which is not supported. A common cause is if there is a package.json file with "type": "module" in the same folder. You can try to fix this by changing the extension of your configuration file to ".cjs" or ".mjs" depending on the content, which will prevent Rollup from trying to preprocess the file but rather hand it to Node directly.`
 	};
 }
 
