@@ -1,5 +1,5 @@
 module.exports = {
 	description: 'uses ES6 module config file',
-	command: 'rollup --config rollup.config.js',
+	command: 'rollup --config rollup.config.js --bundleConfigAsCjs',
 	execute: true
 };
