@@ -1,7 +1,7 @@
 assert.deepEqual(COMMAND_OPTIONS, {
 	_: [],
-	config: 'rollup.config.js',
-	c: 'rollup.config.js',
+	config: 'rollup.config.mjs',
+	c: 'rollup.config.mjs',
 	'some-option': 'foo',
 	'another-option': 42
 });
