@@ -1,6 +1,6 @@
 assert.deepEqual(COMMAND_OPTIONS, {
 	_: [],
-	config: 'rollup.config.js',
-	c: 'rollup.config.js',
+	config: 'rollup.config.mjs',
+	c: 'rollup.config.mjs',
 	silent: true
 });
