@@ -1,4 +1,4 @@
 module.exports = {
 	description: 'correctly generates sourcemaps for multiple outputs',
-	command: 'rollup -c'
+	command: 'rollup -c --bundleConfigAsCjs'
 };
