@@ -1,4 +1,4 @@
 module.exports = {
 	description: 'allows plugins to set options.entry',
-	command: 'rollup -c'
+	command: 'rollup -c --bundleConfigAsCjs'
 };
