@@ -1,5 +1,5 @@
 module.exports = {
 	description: 'uses config file which returns a Promise',
-	command: 'rollup --config rollup.config.js',
+	command: 'rollup --config rollup.config.mjs',
 	execute: true
 };
