@@ -1,5 +1,5 @@
 module.exports = {
 	description: 'defaults to rollup.config.js',
-	command: 'rollup -c',
+	command: 'rollup -c --bundleConfigAsCjs',
 	execute: true
 };
