@@ -1,5 +1,5 @@
 module.exports = {
 	description: 'loads a symlinked config file with the given name',
-	command: 'rollup --config my.rollup.config.js',
+	command: 'rollup --config my.rollup.config.js --bundleConfigAsCjs',
 	execute: true
 };
