@@ -6,7 +6,7 @@ let mainFile;
 
 module.exports = {
 	description: 'recovers from errors during bundling',
-	command: 'rollup -cw',
+	command: 'rollup -cw --bundleConfigAsCjs',
 	before() {
 		mainFile = path.resolve(__dirname, 'main.js');
 		writeFileSync(mainFile, '<=>');
