@@ -1,8 +1,8 @@
 const { mkdirSync, unlinkSync } = require('fs');
 const path = require('path');
-const { writeAndSync, writeAndRetry } = require('../../../../utils');
+const { wait, writeAndSync, writeAndRetry } = require('../../../../utils');
 
-const configFile = path.join(__dirname, 'rollup.config.js');
+const configFile = path.join(__dirname, 'rollup.config.mjs');
 let stopUpdate;
 
 module.exports = {
@@ -49,9 +49,10 @@ module.exports = {
 	},
 	abortOnStderr(data) {
 		if (data === 'initial\n') {
-			stopUpdate = writeAndRetry(
-				configFile,
-				`
+			wait(200).then(() => {
+				stopUpdate = writeAndRetry(
+					configFile,
+					`
 				console.error('updated');
 		    export default {
           input: 'main.js',
@@ -60,7 +61,8 @@ module.exports = {
 		        format: "es"
 		      }
 		    };`
-			);
+				);
+			});
 			return false;
 		}
 		if (data.includes(`created _actual/output2.js`)) {
