@@ -8,7 +8,7 @@ module.exports = {
 	description: 'keeps watching the config file in case the config is changed to an invalid state',
 	command: 'rollup -cw',
 	before() {
-		configFile = path.resolve(__dirname, 'rollup.config.js');
+		configFile = path.resolve(__dirname, 'rollup.config.mjs');
 		writeFileSync(
 			configFile,
 			`
