@@ -16,7 +16,7 @@ module.exports = {
 	description: 'does not rebuild if the config file is updated without change',
 	command: 'rollup -cw',
 	before() {
-		configFile = path.resolve(__dirname, 'rollup.config.js');
+		configFile = path.resolve(__dirname, 'rollup.config.mjs');
 		writeFileSync(configFile, configContent);
 	},
 	after() {
