@@ -1,4 +1,3 @@
 module.exports = {
-	description: 'keeps parameter defaults for class fields',
-	minNodeVersion: 12
+	description: 'keeps parameter defaults for class fields'
 };
